Move generic REST client out of __init__.py
